refactor: Rules 가 가지고 있던 상수들을 연관이 있는 도메인으로 위치 이동
